refactor: Improve file handling and chat sorting
- Update fileUtils.ts to handle file uploads more efficiently
- Modify inputUtils.ts to work with the new file handling approach
- Refactor ChatInput.svelte to use updated file and input utilities
- Implement sortChats function in chatUtils.ts
- Update +page.svelte to use the new sortChats function for better performance
- Adjust Chat.svelte for consistency with new changes

This refactoring improves code organization, reduces duplication, and enhances the overall structure of the chat application.
